Actualizacion

+++ app/src/main/java/com/example/rusbellgutierrez/SRT/Clase_Articulo.java
@@ -2,6 +2,8 @@ package com.example.rusbellgutierrez.SRT;
 
 import android.content.ContentValues;
 
+import java.math.BigInteger;
+
 import static com.example.rusbellgutierrez.SRT.SQL_Columnas.ArticuloEntry;
 
 /**
@@ -10,21 +12,21 @@ import static com.example.rusbellgutierrez.SRT.SQL_Columnas.ArticuloEntry;
 
 public class Clase_Articulo {
 
-    private int idarticulo;
+    private BigInteger idarticulo;
     private String nombre;
-    private int codbarra;
+    private BigInteger codbarra;
 
-    public Clase_Articulo(int idarticulo, String nombre, int codbarra) {
+    public Clase_Articulo(BigInteger idarticulo, String nombre, BigInteger codbarra) {
         this.idarticulo = idarticulo;
         this.nombre = nombre;
         this.codbarra = codbarra;
     }
 
-    public int getIdarticulo() {
+    public BigInteger getIdarticulo() {
         return idarticulo;
     }
 
-    public void setIdarticulo(int idarticulo) {
+    public void setIdarticulo(BigInteger idarticulo) {
         this.idarticulo = idarticulo;
     }
 
@@ -36,11 +38,11 @@ public class Clase_Articulo {
         this.nombre = nombre;
     }
 
-    public int getCodbarra() {
+    public BigInteger getCodbarra() {
         return codbarra;
     }
 
-    public void setCod_barra(int codbarra) {
+    public void setCod_barra(BigInteger codbarra) {
         this.codbarra = codbarra;
     }
 }
